+++ playback/src/audio_backend/jackaudio.rs
@@ -2,7 +2,7 @@ use super::{Open, Sink};
 use crate::config::AudioFormat;
 use crate::convert::Converter;
 use crate::decoder::AudioPacket;
-use crate::player::NUM_CHANNELS;
+use crate::NUM_CHANNELS;
 use jack::{
     AsyncClient, AudioOut, Client, ClientOptions, Control, Port, ProcessHandler, ProcessScope,
 };
@@ -80,4 +80,8 @@ impl Sink for JackSink {
         }
         Ok(())
     }
+}
+
+impl JackSink {
+    pub const NAME: &'static str = "jackaudio";
 }