+++ playback/src/audio_backend/pulseaudio.rs
@@ -2,7 +2,7 @@ use super::{Open, Sink, SinkAsBytes};
 use crate::config::AudioFormat;
 use crate::convert::Converter;
 use crate::decoder::AudioPacket;
-use crate::player::{NUM_CHANNELS, SAMPLE_RATE};
+use crate::{NUM_CHANNELS, SAMPLE_RATE};
 use libpulse_binding::{self as pulse, stream::Direction};
 use libpulse_simple_binding::Simple;
 use std::io;
@@ -55,7 +55,7 @@ impl Sink for PulseAudioSink {
             return Ok(());
         }
 
-        let device = self.device.as_ref().map(|s| (*s).as_str());
+        let device = self.device.as_deref();
         let result = Simple::new(
             None,                // Use the default server.
             APP_NAME,            // Our application's name.
@@ -103,4 +103,8 @@ impl SinkAsBytes for PulseAudioSink {
             ))
         }
     }
+}
+
+impl PulseAudioSink {
+    pub const NAME: &'static str = "pulseaudio";
 }